Re-factored to use ManifestFile object
Realized it would make sense to use OO principles here so I'm
not constantly passing the path and the file itself around a lot.
Put the get_manifest_file method into the init of the class, and
made a method (always run from the init) that generates a backup
of the file.